Create README.md
README

GitHub Open Source Project

Project Name: <Project Name>

Description:
This project is an open-source software project hosted on GitHub. It aims to <provide a brief description of the project's purpose and goals>. The project is developed and maintained by a community of contributors who collaborate on improving and expanding the functionality of the software.